fix_5434 Refactor the query handler for the jdbc source connector
