+++ ohara-connector/src/main/scala/oharastream/ohara/connector/jdbc/source/TimestampIncrementQueryHandler.scala
@@ -17,11 +17,11 @@
 package oharastream.ohara.connector.jdbc.source
 import java.sql.Timestamp
 import java.util.Objects
+
 import oharastream.ohara.client.database.DatabaseClient
 import oharastream.ohara.common.data.{Column, DataType}
 import oharastream.ohara.common.setting.TopicKey
-import oharastream.ohara.common.util.{CommonUtils, Releasable}
-import oharastream.ohara.connector.jdbc.DatabaseProductName.ORACLE
+import oharastream.ohara.common.util.Releasable
 import oharastream.ohara.connector.jdbc.datatype.RDBDataTypeConverterFactory
 import oharastream.ohara.connector.jdbc.util.DateTimeUtils
 import oharastream.ohara.kafka.connector.{RowSourceContext, RowSourceRecord}
@@ -77,14 +77,12 @@ object TimestampIncrementQueryHandler {
       override val rowSourceContext: RowSourceContext = Builder.this.rowSourceContext
       override val topics: Seq[TopicKey]              = Builder.this.topics
       override val schema: Seq[Column]                = Builder.this.schema
-
-      private[this] val client: DatabaseClient = DatabaseClient.builder
+      override protected val client: DatabaseClient = DatabaseClient.builder
         .url(config.dbURL)
         .user(config.dbUserName)
         .password(config.dbPassword)
         .build
       client.connection.setAutoCommit(false)
-      private[this] val dbProduct: String = client.connection.getMetaData.getDatabaseProductName
 
       override protected[source] def queryData(
         key: String,
@@ -192,40 +190,6 @@ object TimestampIncrementQueryHandler {
           client.connection.commit()
         }
       }
-
-      override protected[source] def tableFirstTimestampValue(timestampColumnName: String): Timestamp = {
-        val sql = dbProduct.toUpperCase match {
-          case ORACLE.name =>
-            s"SELECT $timestampColumnName FROM ${config.dbTableName} ORDER BY $timestampColumnName FETCH FIRST 1 ROWS ONLY"
-          case _ =>
-            s"SELECT $timestampColumnName FROM ${config.dbTableName} ORDER BY $timestampColumnName LIMIT 1"
-        }
-
-        val preparedStatement = client.connection.prepareStatement(sql)
-        try {
-          val resultSet = preparedStatement.executeQuery()
-          try {
-            if (resultSet.next()) resultSet.getTimestamp(timestampColumnName)
-            else new Timestamp(CommonUtils.current())
-          } finally Releasable.close(resultSet)
-        } finally Releasable.close(preparedStatement)
-      }
-
-      override protected[source] def current(): Timestamp = {
-        val query = dbProduct.toUpperCase match {
-          case ORACLE.name => "SELECT CURRENT_TIMESTAMP FROM dual"
-          case _           => "SELECT CURRENT_TIMESTAMP;"
-        }
-        val stmt = client.connection.createStatement()
-        try {
-          val rs = stmt.executeQuery(query)
-          try {
-            if (rs.next()) rs.getTimestamp(1) else new Timestamp(0)
-          } finally Releasable.close(rs)
-        } finally Releasable.close(stmt)
-      }
-
-      override def close(): Unit = Releasable.close(client)
     }
   }
 }
